Reshuffle stuff a little
